+++ src/routes/PatronPreRegistrationRecordsDuplicatesPage.js
@@ -1,0 +1,9 @@
+import { PatronPreRegistrationRecordsDuplicates } from '../views';
+
+const PatronPreRegistrationRecordsDuplicatesPage = () => {
+  return (
+    <PatronPreRegistrationRecordsDuplicates />
+  );
+};
+
+export default PatronPreRegistrationRecordsDuplicatesPage;