+++ src/views/PatronsPreRegistrationListContainer/hooks/index.js
@@ -1,0 +1,2 @@
+export { default as useNewRecordHandler } from './useNewRecordHandler';
+export { default as useUserDuplicatesCheck } from './useUserDuplicatesCheck';